Clean up test code
